requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
python-multipart==0.0.6
pillow==10.1.0
python-dotenv==1.0.0
PyMySQL==1.1.0
Werkzeug==2.3.7
cryptography>=41.0.3
py_eureka_client==0.11.0